Multisig wallet features, BCUR, psbt validation, native tpub support, and more
Thanks @jimmysong for the code review on a very similar version of this.

My branch had gotten so unwieldy that I wanted to get it merged as-is. It's much better than the previous version (including a number of small fixes) I and it adds test coverage.

Happy to make any other changes you'd like if you see other issues.

I'll keep my future branches much smaller and nicely organized!